# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ByteShare API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Classrooms
  /classrooms:
    get:
      operationId: getMyClassrooms
      tags: [classrooms]
      summary: Get classrooms I belong to
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of classrooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Classroom"
    post:
      operationId: createClassroom
      tags: [classrooms]
      summary: Create a classroom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomInput"
      responses:
        "201":
          description: Classroom created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
  /classrooms/join:
    post:
      operationId: joinClassroom
      tags: [classrooms]
      summary: Join a classroom by invite code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinClassroomInput"
      responses:
        "200":
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
        "404":
          description: Invalid invite code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classrooms/{classroomId}:
    get:
      operationId: getClassroom
      tags: [classrooms]
      summary: Get classroom by ID
      security:
        - bearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Classroom details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
    patch:
      operationId: updateClassroom
      tags: [classrooms]
      summary: Update classroom (host only)
      security:
        - bearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomUpdate"
      responses:
        "200":
          description: Updated classroom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
  /classrooms/{classroomId}/members:
    get:
      operationId: getClassroomMembers
      tags: [classrooms]
      summary: Get classroom members
      security:
        - bearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
  /classrooms/{classroomId}/members/{memberId}:
    delete:
      operationId: removeMember
      tags: [classrooms]
      summary: Remove a member (host only)
      security:
        - bearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /classrooms/{classroomId}/stats:
    get:
      operationId: getClassroomStats
      tags: [classrooms]
      summary: Get classroom statistics (host only)
      security:
        - bearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Classroom stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomStats"

  # Resources
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # Upload signature
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, tokenBalance, reputation]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        tokenBalance:
          type: integer
        reputation:
          type: integer
        createdAt:
          type: string
          format: date-time
    ClassroomInput:
      type: object
      required: [name, subject, department, semester]
      properties:
        name:
          type: string
          minLength: 3
        subject:
          type: string
        department:
          type: string
        semester:
          type: string
        description:
          type: ["string", "null"]
    ClassroomUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    JoinClassroomInput:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
          minLength: 6
          maxLength: 6
    Classroom:
      type: object
      required: [id, name, subject, department, semester, inviteCode, hostId, memberCount]
      properties:
        id:
          type: string
        name:
          type: string
        subject:
          type: string
        department:
          type: string
        semester:
          type: string
        description:
          type: ["string", "null"]
        inviteCode:
          type: string
        hostId:
          type: string
        hostName:
          type: string
        memberCount:
          type: integer
        isHost:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Member:
      type: object
      required: [id, name, email, reputation, joinedAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        reputation:
          type: integer
        joinedAt:
          type: string
          format: date-time
    ClassroomStats:
      type: object
      required: [totalMembers, totalApproved, totalPending, totalTokensAwarded, topContributors]
      properties:
        totalMembers:
          type: integer
        totalApproved:
          type: integer
        totalPending:
          type: integer
        totalTokensAwarded:
          type: integer
        topContributors:
          type: array
          items:
            $ref: "#/components/schemas/Contributor"
    Contributor:
      type: object
      required: [id, name, reputation, approvedCount]
      properties:
        id:
          type: string
        name:
          type: string
        reputation:
          type: integer
        approvedCount:
          type: integer
...
    Notification:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [resource_approved, resource_rejected, removed_from_classroom]
        message:
          type: string
        isRead:
          type: boolean
        resourceTitle:
          type: ["string", "null"]
        classroomName:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...